Revert "memory issue fix"
This reverts commit 0882cdd0b738d0c8b7b4e3e00855d746933f4e39.

+++ sauce/game/gamesession.h
@@ -48,7 +48,7 @@ typedef struct GameSession {
     Boolean colorSupport; // Support for color screens
 
     Button *displayButtons;
-    int displayButtonCount;
+    UInt8 displayButtonCount;
 
     Movement *movement;
     AttackAnimation *attackAnimation;
