+++ sauce/game/movement.h
@@ -5,8 +5,8 @@
 
 int movement_distance(Coordinate axialA, Coordinate axialB);
 Coordinate movement_coordinateAtPercentageOfLine(Line line, float percentage);
-Coordinate movement_coordinateAtPercentageOfTrajectory(Trajectory trajectory, float percentage, int *orientation);
+Coordinate movement_coordinateAtPercentageOfTrajectory(Trajectory trajectory, float percentage, UInt8 *orientation);
 Trajectory movement_trajectoryBetween(Coordinate startCoordinate, Coordinate endCoordinate);
 void movement_findTilesInRange(Coordinate currentPosition, int maxTileRange, Coordinate *invalidCoordinates, int invalidCoordinatesCount, Coordinate **results, int *numberOfPositions);
-int movement_orientationBetween(Coordinate coordA, Coordinate coordB);
+UInt8 movement_orientationBetween(Coordinate coordA, Coordinate coordB);
 #endif
